pyproject.toml
[tool.poetry]
name = "drive-inventory"
version = "0.1.0"
description = ""
authors = ["Mário Idival <marioidival@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.129.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
